requirements.txt
selenium
beautifulsoup4
requests
pandas
python-dotenv
PyGithub
